Added Expense entity and repository

+++ src/main/java/se/cygni/expenses/api/Event.java
@@ -4,20 +4,20 @@ import java.util.Date;
 
 public class Event {
 
-    private int id;
+    private long id;
     private String name;
     private Date date;
 
     public Event() {
     }
 
-    public Event(int id, String name, Date date) {
+    public Event(long id, String name, Date date) {
         this.id = id;
         this.name = name;
         this.date = date;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
